issue-5894: renamed mirror_unsafe / MirrorUnsafe -> naive_mirrored / NaiveMirrored
