+++ cloud/filestore/libs/storage/fastshard/impl/naive_mirrored/shard.h
@@ -14,7 +14,7 @@ namespace NCloud::NFileStore::NStorage::NFastShard {
 
 ////////////////////////////////////////////////////////////////////////////////
 
-IFileSystemShardPtr CreateMirrorUnsafeFileSystemShard(
+IFileSystemShardPtr CreateNaiveMirroredFileSystemShard(
     ui32 shardNo,
     const NProtoPrivate::TPersistentFastShardConfig& config);
