+++ cloud/filestore/libs/storage/fastshard/impl/naive_mirrored/ut/ya.make
@@ -5,7 +5,7 @@ SRCS(
 )
 
 PEERDIR(
-    cloud/filestore/libs/storage/fastshard/impl/mirror_unsafe
+    cloud/filestore/libs/storage/fastshard/impl/naive_mirrored
     cloud/filestore/libs/storage/fastshard/sn/impl
     cloud/filestore/libs/storage/fastshard/sn/server
     cloud/filestore/libs/storage/fastshard/sn/quorum
